refactors filesystem config to add with local files mixin
